+++ src/custom/WidgetPicker/index.tsx
@@ -1,1 +1,1 @@
-export * from './WidgetPicker';
+export { WidgetPicker, type WidgetPickerProps, type WidgetItem } from './WidgetPicker';
